rename view package
